aggregate site with reference collection

+++ app/controllers/sites_controller.rb
@@ -56,15 +56,14 @@ class SitesController < ApplicationController
     site_params = JSON.parse params[:site]
     ui_attributes = prepare_from_ui(site_params)
     site = collection.sites.new(ui_attributes.merge(user: current_user))
-    site.user_id = current_user.id
-    if site.valid?
-      site.save!
-      current_user.site_count += 1
-      current_user.update_successful_outcome_status
-      current_user.save!(:validate => false)
-      render json: site, :layout => false
+    site.user = current_user
+
+    site_aggregator = SiteAggregator.new(site)
+
+    if site_aggregator.save
+      render json: site_aggregator.site, :layout => false
     else
-      render json: site.errors.messages, status: :unprocessable_entity, :layout => false
+      render json: site_aggregator.site.errors.messages, status: :unprocessable_entity, :layout => false
     end
   end
 
@@ -73,14 +72,13 @@ class SitesController < ApplicationController
     site.user = current_user
     site.properties_will_change!
     site.attributes = prepare_from_ui(site_params)
-    if site.valid?
-      site.save!
-      if params[:photosToRemove]
-        Site::UploadUtils.purgePhotos(params[:photosToRemove])
-      end
-      render json: site, :layout => false
+
+    site_aggregator = SiteAggregator.new(site)
+    if site_aggregator.save
+      Site::UploadUtils.purgePhotos(params[:photosToRemove]) if params[:photosToRemove]
+      render json: site_aggregator.site, :layout => false
     else
-      render json: site.errors.messages, status: :unprocessable_entity, :layout => false
+      render json: site_aggregator.site.errors.messages, status: :unprocessable_entity, :layout => false
     end
   end
 
@@ -94,11 +92,12 @@ class SitesController < ApplicationController
     site.properties_will_change!
 
     site.properties[params[:es_code]] = field.decode_from_ui(params[:value])
-    if site.valid?
-      site.save!
-      render json: site, :status => 200, :layout => false
+    site_aggregator = SiteAggregator.new(site)
+
+    if site_aggregator.save
+      render json: site_aggregator.site, :status => 200, :layout => false
     else
-      error_message = site.errors[:properties][0][params[:es_code]]
+      error_message = site_aggregator.site.errors[:properties][0][params[:es_code]]
       render json: {:error_message => error_message}, status: :unprocessable_entity, :layout => false
     end
   end
@@ -108,7 +107,7 @@ class SitesController < ApplicationController
   #   search = MapSearch.new params[:collection_ids], user: current_user
 
   #   formula = params[:formula].downcase if params[:formula].present?
-    
+
   #   search.set_formula formula if formula.present?
   #   search.zoom = zoom
   #   search.bounds = params if zoom >= 2
@@ -122,9 +121,9 @@ class SitesController < ApplicationController
   #   end
 
   #   search.where params.except(:action, :controller, :format, :n, :s, :e, :w, :z, :collection_ids, :exclude_id, :search, :hierarchy_code, :selected_hierarchies, :_alert, :formula)
-    
+
   #   search.prepare_filter
-  #   # search.apply_queries 
+  #   # search.apply_queries
 
   #   render json: search.results
   # end
@@ -148,7 +147,7 @@ class SitesController < ApplicationController
     search = MapSearch.new [collection_id], user: current_user
 
     formula = params[:formula].downcase if params[:formula].present?
-    
+
     search.set_formula formula if formula.present?
     search.zoom = zoom
     search.bounds = params if zoom >= 2
@@ -162,9 +161,9 @@ class SitesController < ApplicationController
     end
 
     search.where params.except(:action, :controller, :format, :n, :s, :e, :w, :z, :collection_ids, :exclude_id, :search, :hierarchy_code, :selected_hierarchies, :_alert, :formula)
-    
+
     search.prepare_filter
-    # search.apply_queries 
+    # search.apply_queries
 
     return search.results
   end
@@ -175,7 +174,7 @@ class SitesController < ApplicationController
   #   search = MapSearch.new params[:collection_ids], user: current_user
 
   #   formula = params[:formula].downcase if params[:formula].present?
-    
+
   #   search.set_formula formula if formula.present?
   #   search.zoom = zoom
   #   search.bounds = params if zoom >= 2
@@ -187,10 +186,10 @@ class SitesController < ApplicationController
   #     search.selected_hierarchy params[:hierarchy_code], params[:selected_hierarchies]
   #   end
   #   search.where params.except(:action, :controller, :format, :n, :s, :e, :w, :z, :collection_ids, :exclude_id, :search, :hierarchy_code, :selected_hierarchies, :_alert, :formula)
-    
+
   #   # search.apply_queries
   #   search.prepare_filter
-  #   render json: search.sites_json    
+  #   render json: search.sites_json
   # end
 
   def search_alert_site
@@ -212,7 +211,7 @@ class SitesController < ApplicationController
     search = MapSearch.new [collection_id], user: current_user
 
     formula = params[:formula].downcase if params[:formula].present?
-    
+
     search.set_formula formula if formula.present?
     search.zoom = zoom
     search.bounds = params if zoom >= 2
@@ -224,10 +223,10 @@ class SitesController < ApplicationController
       search.selected_hierarchy params[:hierarchy_code], params[:selected_hierarchies]
     end
     search.where params.except(:action, :controller, :format, :n, :s, :e, :w, :z, :collection_ids, :exclude_id, :search, :hierarchy_code, :selected_hierarchies, :_alert, :formula)
-    
+
     # search.apply_queries
     search.prepare_filter
-    return search.sites_json    
+    return search.sites_json
   end
 
   def destroy
@@ -307,7 +306,7 @@ class SitesController < ApplicationController
     decoded_properties = {}
     site_properties = parameters.delete("properties") || {}
     files = params[:fileUpload] || {}
-    
+
     site_properties.each_pair do |es_code, value|
       value = [ value, files[value] ] if fields[es_code].kind == 'photo'
       decoded_properties[es_code] = fields[es_code].decode_from_ui(value)
